added clarity table, fixed colors, added service
