+++ src/app/progress/progress-card/progress-card.component.scss
@@ -1,19 +1,3 @@
-.status-provisioning::-moz-progress-bar{
-    background-color:  var(--clr-color-warning-100, yellow);
-}
-
-.status-running::-moz-progress-bar{
-    background-color:  var(--clr-progress-default-color,#54cbf2);
-}
-
-.status-finished::-moz-progress-bar{
-    background-color:  var(--clr-color-danger-100,red);
-}
-
-.status-success::-moz-progress-bar{
-    background-color:  var(--clr-color-success-100,green);
-}
-
 .progress span {
     margin-top: -0.5em;
 }
