+++ src/styles.scss
@@ -104,3 +104,19 @@ overflows the modal and hence displays an unneeded scrollbar */
   margin: 0px;
   color: var(--clr-tooltip-color, #fff);
 }
+
+.status-provisioning{
+  --clr-progress-default-color:  var(--clr-color-warning-100, yellow);
+}
+
+.status-running{
+  --clr-progress-default-color:  var(--clr-progress-default-color,#54cbf2);
+}
+
+.status-finished{ 
+  --clr-progress-default-color:  var(--clr-color-danger-100,red);
+}
+
+.status-success{
+  --clr-progress-default-color:  var(--clr-color-success-100,green);
+}
